pyproject.toml
[project]
name = "qq-plotter"
version = "0.1.0"
description = "This package makes qq-polts based on specified data and distribution type"
authors = [
    {name = "Nikita Zuev"}
]
readme = "README.md"
requires-python = "^3.12"

[tool.poetry.scripts]
main = "src.main:main"

[tool.poetry.dependencies]
pyyaml = "^6.0.2"
loguru = "^0.7.3"
python-dotenv = "^1.1.0"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.2.0"
mypy = "^1.15.0"

[tool.black]
line-length = 79
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 79
src_paths = ["src", "tests"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
disallow_untyped_defs = true
ignore_missing_imports = true
exclude = "migrations|.venv"

[tool.poetry.group.math.dependencies]
matplotlib = "^3.10.3"
scipy = "^1.15.3"
numpy = "^2.2.5"
pandas = "^2.2.3"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
